weights update added
The weights update is added using an approximation method found in XR-EE-SB 2012:015, as validated by an interpretation of 978-1580536318/158053631X.

The measurement prediction has been moved before the state update since the  time frame that the prediction is made should usually be before that of which the prediction pertains to.

The sensor data has also been explicitly returned instead of altering them globally.